% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/amomoStan.R
\name{amomoStan}
\alias{amomoStan}
\title{A bayesian version of the A-MOMO algorithm}
\usage{
amomoStan(data, spring = 16:25, autumn = 31:47, datevar = "date",
  mortvar = "n", popvar = NA, byvar = NA)
}
\arguments{
\item{data}{a data frame with all necessary variables}

\item{spring}{numbers of the weeks defining Spring}

\item{autumn}{numbers of the weeks defining Autumn}

\item{datevar}{name of the variable containing the date of the observation (preferrably the date of the Monday)}

\item{mortvar}{name of the variable containing the number of deaths}

\item{popvar}{(optionally) the name of the variable containing the population denominator}

\item{byvar}{(optionally) the name of the variable by which groups the analysis is done}
}
\value{
stan fit object
}
\description{
Fits a Serfling -type model to a mortality time series by omiting summer and winter seasons from the estimation, and making predictions for the whole time span of the data. This version does not include delay adjustments
}
