% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flumomoStan.R
\name{flumomoStan}
\alias{flumomoStan}
\title{A bayesian version of the FluMOMO algorithm}
\usage{
flumomoStan(data, spring = 16:25, autumn = 31:47, datevar = "date",
  mortvar = "n", covar = NULL, seasvar = NULL, popvar = NA,
  byvar = NA, penalties = NULL, data.only = FALSE)
}
\arguments{
\item{data}{a data frame with all necessary variables}

\item{datevar}{name of the variable containing the date of the observation (preferrably the date of the Monday)}

\item{mortvar}{name of the variable containing the number of deaths}

\item{covar}{character vector of covariates included in the model as such}

\item{seasvar}{character vector of covariates included in the model by season}

\item{popvar}{(optionally) the name of the variable containing the population denominator}

\item{byvar}{(optionally) the name of the variable by which groups the analysis is done}

\item{penalties}{control the shrinkage between}

\item{data.only}{if TRUE just collect all the data together and don't fit}
}
\value{
stan fit object
}
\description{
Fits a regression model to a mortality time series using any number of covariates with effects constrained to be positive.
}
