% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{momoCalendar}
\alias{momoCalendar}
\title{A calendar of weeks}
\usage{
momoCalendar(start, end)
}
\arguments{
\item{start}{first week to consider (a Date rounded to a Monday)}

\item{end}{last week to consider (a Date rounded to a Monday)}
}
\value{
a data frame with fields date ...
}
\description{
Create a dataset of weeks with additional metadata. Useful for extending the observed data to incorporate future dates for forecasting
}
