% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{momoSeason}
\alias{momoSeason}
\title{Calculate the season from a date}
\usage{
momoSeason(date, start = 26, fact = FALSE)
}
\arguments{
\item{date}{sate to use}

\item{start}{the week when year starts}

\item{fact}{if TRUE, return a factor, otherwise return a number}
}
\value{
number giving the start year of the season or a factor
}
\description{
Define season as a year that starts given week. If this week is 1, season corresponds to a normal year
}
