% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{momoSin}
\alias{momoSin}
\title{Sinusoidal covariates}
\usage{
momoSin(date, n = 1, freq = NA, use = NULL)
}
\arguments{
\item{date}{a Date variable to base the waves on}

\item{n}{number of wavelengths}

\item{freq}{base frequency of the longest wave in years. Default 1 means longest wave is yearly}

\item{use}{which frequencies to use. Default is 1 to n}
}
\value{
a matrix of components
}
\description{
Create n x 2 covariates with varying wave lengths
}
