\name{momoStan-package}
\alias{momoStan-package}
\alias{momoStan}
\docType{package}
\title{
\packageTitle{momoStan}
}
\description{
\packageDescription{momoStan}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{momoStan}
\packageIndices{momoStan}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{momoStan}

Maintainer: \packageMaintainer{momoStan}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
}
