% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{momoTrend}
\alias{momoTrend}
\title{A trend effect from a date}
\usage{
momoTrend(date, origin = NULL)
}
\arguments{
\item{date}{a Date variable}

\item{origin}{origin date, defaults to the midpoint}
}
\value{
a numeric vector with values between -1 and 1
}
\description{
Turns a Date variable into a trend suitable for modeling. Currently only linear trend is available
}
