% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{na.0}
\alias{na.0}
\title{Replace missing values with zeroes}
\usage{
na.0(a, b = 0)
}
\arguments{
\item{a}{variable to be transformed}

\item{b}{value that used to replace missing values, defaults 0. Recycled}
}
\value{
transformed variable
}
\description{
Replace missing values with zeroes
}
