% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{plot.momoStanEffs}
\alias{plot.momoStanEffs}
\title{Plot method for momoStanEffs}
\usage{
\method{plot}{momoStanEffs}(obj, type = "baseline", colour = "red",
  group = 1, alpha = 0.2)
}
\arguments{
\item{obj}{an object to plot}

\item{type}{type of the plot}

\item{colour}{colour of the ribbon}

\item{group}{which group to be plotted}

\item{alpha}{transparency of the ribbon}
}
\value{
ggplot
}
\description{
Plot method for momoStanEffs
}
