% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{momoEffects}
\alias{momoEffects}
\title{Extract interesting information from a fit}
\usage{
momoEffects(obj, temp = "TEMP", infl = "InfA")
}
\arguments{
\item{obj}{fit object}

\item{temp}{either numeric vector or a regular expression to identify temperature effects}

\item{infl}{either numeric vector or a regular expression to identify influenxa effects}
}
\value{
a list with components
}
\description{
Extract interesting information from a fit
}
